# docs/apifox/zhiguang-openapi.yaml
openapi: 3.0.3
info:
  title: 知光 Zhiguang 前后端联调 API
  version: "1.0.0"
  description: |
    本文档根据当前后端源码整理，面向本地前后端联调与 Apifox 导入。

    本地默认后端地址为 `http://localhost:8080`。除特别标注为公开接口外，接口都需要在请求头中携带：

    `Authorization: Bearer <accessToken>`

    认证流程建议：

    1. 调用“发送验证码”。
    2. 本地开发环境从后端日志或 Redis 读取验证码。
    3. 调用“注册”或“登录”，拿到 `accessToken` 和 `refreshToken`。
    4. 在 Apifox 环境变量中配置 `accessToken`、`refreshToken`、`postId`、`targetUserId` 等变量。
servers:
  - url: http://localhost:8080
    description: 本地 Spring Boot 后端
...
paths:
...
  /api/v1/knowposts/drafts:
    post:
      tags:
        - KnowPost
      summary: 创建知文草稿
      description: 创建空草稿，默认类型为 `image_text`，默认可见性为 `public`，返回新草稿 ID。
      operationId: createKnowPostDraft
      security:
        - bearerAuth: []
      responses:
        "200":
          description: 草稿创建成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KnowPostDraftCreateResponse"
              examples:
                success:
                  value:
                    id: "1950000000000000000"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/knowposts/{id}/content/confirm:
    post:
      tags:
        - KnowPost
      summary: 确认知文正文上传
      description: 前端通过 OSS 预签名 URL 上传正文后，回传对象存储信息，后端保存 `contentUrl`。
      operationId: confirmKnowPostContent
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/KnowPostId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KnowPostContentConfirmRequest"
            examples:
              markdown:
                value:
                  objectKey: posts/1950000000000000000/content.md
                  etag: "9b2cf535f27731c974343645a3985328"
                  size: 2048
                  sha256: b94d27b9934d3e08a52e52d7da7dabfadeadbeef1234567890abcdef12345678
      responses:
        "204":
          description: 确认成功，无响应体
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/knowposts/{id}:
    patch:
      tags:
        - KnowPost
      summary: 更新知文元数据
      description: |
        局部更新标题、标签、封面图片、可见性、置顶状态、简介等元数据。

        `visible` 支持：`public`、`followers`、`school`、`private`、`unlisted`。
      operationId: patchKnowPostMetadata
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/KnowPostId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KnowPostPatchRequest"
            examples:
              full:
                value:
                  title: Spring Security JWT 认证链路总结
                  tagId: 1
                  tags:
                    - Java
                    - Spring Security
                    - JWT
                  imgUrls:
                    - https://cdn.zhiguang.cn/posts/1950000000000000000/images/cover.png
                  visible: public
                  isTop: false
                  description: 从过滤链到 Resource Server 的完整认证链路
              partial:
                value:
                  title: 更新后的标题
                  description: 更新后的简介
      responses:
        "204":
          description: 更新成功，无响应体
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
    delete:
      tags:
        - KnowPost
      summary: 删除知文
      description: 软删除知文，仅作者本人可操作。
      operationId: deleteKnowPost
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/KnowPostId"
      responses:
        "204":
          description: 删除成功，无响应体
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/knowposts/{id}/publish:
    post:
      tags:
        - KnowPost
      summary: 发布知文
      description: 将知文置为 `published`，并通过 Outbox 异步创建 RAG 索引；向量失败不会回滚发布。
      operationId: publishKnowPost
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/KnowPostId"
      responses:
        "200":
          description: 发布成功，RAG 索引进入排队状态
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KnowPostPublishResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/knowposts/{id}/top:
    patch:
      tags:
        - KnowPost
      summary: 设置知文置顶状态
      operationId: patchKnowPostTop
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/KnowPostId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KnowPostTopPatchRequest"
            examples:
              top:
                value:
                  isTop: true
              cancelTop:
                value:
                  isTop: false
      responses:
        "204":
          description: 更新成功，无响应体
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/knowposts/{id}/visibility:
    patch:
      tags:
        - KnowPost
      summary: 设置知文可见性
      operationId: patchKnowPostVisibility
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/KnowPostId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KnowPostVisibilityPatchRequest"
            examples:
              public:
                value:
                  visible: public
              private:
                value:
                  visible: private
      responses:
        "204":
          description: 更新成功，无响应体
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/knowposts/feed:
    get:
      tags:
        - KnowPost
      summary: 首页公开 Feed
      description: |
        公开接口。查询公开、已发布知文。可不带 token；如果携带 Bearer token，返回项中的 `liked`、`faved` 可体现当前用户状态。
      operationId: getPublicFeed
      security:
        - {}
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
          example: 1
        - name: size
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 20
          example: 20
      responses:
        "200":
          description: Feed 分页结果
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeedPageResponse"
              examples:
                list:
                  $ref: "#/components/examples/FeedPage"
  /api/v1/knowposts/mine:
    get:
      tags:
        - KnowPost
      summary: 我的已发布知文
      operationId: getMyKnowPosts
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
          example: 1
        - name: size
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 20
          example: 20
      responses:
        "200":
          description: 我的知文分页结果
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeedPageResponse"
              examples:
                list:
                  $ref: "#/components/examples/FeedPage"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/knowposts/detail/{id}:
    get:
      tags:
        - KnowPost
      summary: 知文详情
      description: |
        公开接口。公开已发布内容可匿名访问；非公开内容需要作者本人身份，服务层会做权限判断。
      operationId: getKnowPostDetail
      security:
        - {}
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/KnowPostId"
      responses:
        "200":
          description: 知文详情
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KnowPostDetailResponse"
              examples:
                detail:
                  $ref: "#/components/examples/KnowPostDetail"
        "400":
          $ref: "#/components/responses/BadRequest"
...
  /api/v1/knowposts/{id}/qa/stream:
    get:
      tags:
        - AI/RAG
      summary: 单篇知文 RAG 流式问答
      description: |
        公开接口，返回 `text/event-stream`。后端会确保该知文被索引，然后基于向量召回上下文进行流式回答。
      operationId: streamKnowPostQa
      security: []
      parameters:
        - $ref: "#/components/parameters/KnowPostId"
        - name: question
          in: query
          required: true
          schema:
            type: string
          example: 这篇文章里 Refresh Token 为什么要放 Redis？
        - name: topK
          in: query
          required: false
          schema:
            type: integer
            default: 5
            minimum: 1
          example: 5
        - name: maxTokens
          in: query
          required: false
          schema:
            type: integer
            default: 1024
            minimum: 1
          example: 1024
      responses:
        "200":
          description: SSE 流式回答
          content:
            text/event-stream:
              schema:
                type: string
              examples:
                stream:
                  value: |
                    data: Refresh Token 放 Redis 主要是为了支持服务端撤销和 TTL 管理。

                    data: 这样登出、重置密码、令牌轮换时可以立即让旧 token 失效。
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/v1/knowposts/{id}/rag/reindex:
    post:
      tags:
        - AI/RAG
      summary: 手动重建单篇知文向量索引
      description: 需要鉴权且仅作者可调用。返回结构化索引状态。
      operationId: reindexKnowPostRag
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/KnowPostId"
      responses:
        "200":
          description: 重建成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RagIndexStatusResponse"
              examples:
                success:
                  value:
                    postId: "335705124099985408"
                    status: READY
                    chunkCount: 47
                    indexedAt: "2026-07-15T17:50:00+08:00"
                    message: AI问答已就绪
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/knowposts/{id}/rag/status:
    get:
      tags:
        - AI/RAG
      summary: 查询单篇知文 RAG 索引状态
      description: 公开知文允许匿名查询；非公开知文仅作者可查询。
      operationId: getKnowPostRagStatus
      parameters:
        - $ref: "#/components/parameters/KnowPostId"
      responses:
        "200":
          description: 当前索引状态
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RagIndexStatusResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 登录或注册成功后返回的 accessToken。
  parameters:
    KnowPostId:
      name: id
      in: path
      required: true
      description: 知文 ID。源码中接收 long，文档示例按字符串展示以避免前端大整数精度问题。
      schema:
        type: string
        pattern: "^\\d+$"
      example: "{{postId}}"
...
  responses:
    BadRequest:
      description: 业务异常或参数校验失败
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            badRequest:
              value:
                code: BAD_REQUEST
                message: 请求参数错误
            identifierExists:
              value:
                code: IDENTIFIER_EXISTS
                message: 账号已存在
            invalidCredentials:
              value:
                code: INVALID_CREDENTIALS
                message: 登录凭证错误
    Unauthorized:
      description: 未认证或 Bearer JWT 无效/过期
  examples:
...
    FeedPage:
      value:
        items:
          - id: "1950000000000000000"
            title: Spring Security JWT 认证链路总结
            description: 从过滤链到 Resource Server 的完整认证链路
            coverImage: https://cdn.zhiguang.cn/posts/1950000000000000000/images/cover.png
            tags:
              - Java
              - JWT
            authorAvatar: https://static.zhiguang.cn/default-avatar.png
            authorNickname: 宋同学
            tagJson: "[\"Java\",\"AI应用\"]"
            likeCount: 12
            favoriteCount: 3
            liked: false
            faved: false
            isTop: false
        page: 1
        size: 20
        hasMore: false
    KnowPostDetail:
      value:
        id: "1950000000000000000"
        title: Spring Security JWT 认证链路总结
        description: 从过滤链到 Resource Server 的完整认证链路
        contentUrl: https://cdn.zhiguang.cn/posts/1950000000000000000/content.md
        images:
          - https://cdn.zhiguang.cn/posts/1950000000000000000/images/cover.png
        tags:
          - Java
          - Spring Security
          - JWT
        authorId: "100"
        authorAvatar: https://static.zhiguang.cn/default-avatar.png
        authorNickname: 宋同学
        authorTagJson: "[\"Java\",\"AI应用\"]"
        likeCount: 12
        favoriteCount: 3
        liked: false
        faved: false
        isTop: false
        visible: public
        type: image_text
        publishTime: "2026-07-12T10:00:00Z"
  schemas:
    ErrorResponse:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          enum:
            - IDENTIFIER_EXISTS
            - IDENTIFIER_NOT_FOUND
            - ZGID_EXISTS
            - VERIFICATION_RATE_LIMIT
            - VERIFICATION_DAILY_LIMIT
            - VERIFICATION_NOT_FOUND
            - VERIFICATION_MISMATCH
            - VERIFICATION_TOO_MANY_ATTEMPTS
            - INVALID_CREDENTIALS
            - PASSWORD_POLICY_VIOLATION
            - TERMS_NOT_ACCEPTED
            - REFRESH_TOKEN_INVALID
            - BAD_REQUEST
            - INTERNAL_ERROR
        message:
          type: string
...
    Visible:
      type: string
      enum:
        - public
        - followers
        - school
        - private
        - unlisted
...
    KnowPostDraftCreateResponse:
      type: object
      required:
        - id
      properties:
        id:
          type: string
          description: 新草稿 ID。
    KnowPostPublishResponse:
      type: object
      required:
        - postId
        - publishStatus
        - ragIndexStatus
        - ragMessage
      properties:
        postId:
          type: string
        publishStatus:
          type: string
          enum: [PUBLISHED]
        ragIndexStatus:
          type: string
          enum: [PENDING]
        ragMessage:
          type: string
    RagIndexStatusResponse:
      type: object
      required:
        - postId
        - status
        - chunkCount
        - message
      properties:
        postId:
          type: string
        status:
          type: string
          enum: [NOT_INDEXED, PENDING, INDEXING, READY, FAILED, SKIPPED]
        chunkCount:
          type: integer
          format: int32
        indexedAt:
          type: string
          format: date-time
          nullable: true
        message:
          type: string
    KnowPostContentConfirmRequest:
      type: object
      required:
        - objectKey
        - etag
        - size
        - sha256
      properties:
        objectKey:
          type: string
        etag:
          type: string
        size:
          type: integer
          format: int64
        sha256:
          type: string
    KnowPostPatchRequest:
      type: object
      properties:
        title:
          type: string
        tagId:
          type: integer
          format: int64
          nullable: true
        tags:
          type: array
          maxItems: 20
          items:
            type: string
        imgUrls:
          type: array
          maxItems: 20
          items:
            type: string
            format: uri
        visible:
          $ref: "#/components/schemas/Visible"
        isTop:
          type: boolean
        description:
          type: string
    KnowPostTopPatchRequest:
      type: object
      required:
        - isTop
      properties:
        isTop:
          type: boolean
    KnowPostVisibilityPatchRequest:
      type: object
      required:
        - visible
      properties:
        visible:
          $ref: "#/components/schemas/Visible"
    FeedItemResponse:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
          nullable: true
        coverImage:
          type: string
          format: uri
          nullable: true
        tags:
          type: array
          items:
            type: string
        authorAvatar:
          type: string
          nullable: true
        authorNickname:
          type: string
          nullable: true
        tagJson:
          type: string
          nullable: true
        likeCount:
          type: integer
          format: int64
        favoriteCount:
          type: integer
          format: int64
        liked:
          type: boolean
          nullable: true
        faved:
          type: boolean
          nullable: true
        isTop:
          type: boolean
          nullable: true
    FeedPageResponse:
      type: object
      required:
        - items
        - page
        - size
        - hasMore
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/FeedItemResponse"
        page:
          type: integer
        size:
          type: integer
        hasMore:
          type: boolean
    KnowPostDetailResponse:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
          nullable: true
        contentUrl:
          type: string
          format: uri
          nullable: true
        images:
          type: array
          items:
            type: string
        tags:
          type: array
          items:
            type: string
        authorId:
          type: string
        authorAvatar:
          type: string
          nullable: true
        authorNickname:
          type: string
          nullable: true
        authorTagJson:
          type: string
          nullable: true
        likeCount:
          type: integer
          format: int64
        favoriteCount:
          type: integer
          format: int64
        liked:
          type: boolean
          nullable: true
        faved:
          type: boolean
          nullable: true
        isTop:
          type: boolean
          nullable: true
        visible:
          $ref: "#/components/schemas/Visible"
        type:
          type: string
          example: image_text
        publishTime:
          type: string
          format: date-time
          nullable: true
...